requirements.txt
Flask==3.1.3
biopython==1.86